api/requirements.txt
asgiref==3.7.2
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==4.2.7
djangorestframework==3.14.0
gunicorn==21.2.0
mpmath==1.3.0
packaging==23.2
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.8.0
